--- a/packages/modeling/src/operations/expansions/expandShell.js
+++ b/packages/modeling/src/operations/expansions/expandShell.js
@@ -29,15 +29,22 @@ const mapPlaneToVertex = (map, vertex, plane) => {
   }
 }
 
+/**
+ * Collect all planes adjacent to each edge.
+ * Combine undirected edges, no need for duplicate cylinders.
+ */
 const mapPlaneToEdge = (map, edge, plane) => {
-  const i = map.findIndex((item) => (vec3.equals(item[0], edge[0]) && vec3.equals(item[1], edge[1])) || (vec3.equals(item[0], edge[1]) && vec3.equals(item[1], edge[0])))
-  if (i < 0) {
+  const key0 = edge[0].toString()
+  const key1 = edge[1].toString()
+  // Sort keys to make edges undirected
+  const key = key0 < key1 ? `${key0},${key1}` : `${key1},${key0}`
+  if (!map.has(key)) {
     const entry = [edge, [plane]]
-    map.push(entry)
-    return
+    map.set(key, entry)
+  } else {
+    const planes = map.get(key)[1]
+    planes.push(plane)
   }
-  const planes = map[i][1]
-  planes.push(plane)
 }
 
 const addUniqueAngle = (map, angle) => {
@@ -65,7 +72,7 @@ const expandShell = (options, geometry) => {
 
   let result = geom3.create()
   const vertices2planes = new Map() // {vertex: [vertex, [plane, ...]]}
-  const edges2planes = [] // contents: [[vertex, vertex], [plane, ...]]
+  const edges2planes = new Map() // {edge: [[vertex, vertex], [plane, ...]]}
 
   const v1 = vec3.create()
   const v2 = vec3.create()