There is a bug in [expandShell.js](https://github.com/jscad/OpenJSCAD.org/blob/b06288af2fe8f7b007e4b10eac413c74bd6c1bfa/packages/modeling/src/operations/expansions/expandShell.js#L31) where the edge check in findIndex is always false, because it is comparing an edge to a vertex.

The result of that bug is that two cylinders are added to the expansion for each edge, instead of just one. Two cylinders on each edge is unnecessary and slows the computation down. This new code produces cleaner geometry too.

I also changed the functions to avoid linear searches for unique vertices and edges. Actually that's where I started, I only noticed the bug after I accidentally fixed it.

The following example (which is a real test from [expand.test.js](https://github.com/jscad/OpenJSCAD.org/blob/c8a37ad6430aec8361a5e121730b34629c0d969c/packages/modeling/src/operations/expansions/expand.test.js#L120-L123)) reduced runtime from 4200ms on master, to 1600ms on this branch. This is a 10% reduction in total `npm test` time! (23s -> 20s)

```js
const jscad = require('@jscad/modeling')
const { expand } = jscad.expansions
const { sphere } = jscad.primitives

const main = () => {
  const obj = sphere({ radius: 5, segments: 8 })
  return expand({ delta: 5 }, obj)
}

module.exports = { main }
```

### All Submissions:

* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../pulls) for the same update/change?
* [x] Does your submission pass tests?