@z3dev 2022-02-23
@platypii can you try this with the examples in #876 ? i'm curious to see if the bug fix helps.